requirements.txt
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.6
channels==3.0.4
constantly==15.1.0
cryptography==3.4.8
daphne==3.0.2
Django==3.2.6
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-simple-math-captcha==1.0.9
django-utils-six==2.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.7.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
zope.interface==5.4.0
